CrimeQuery prompt
+ via DoubleClick on health bar
+ via DoubleClick on Mobile in game
#434
